requirements.txt
discord.py
slack-sdk
SQLAlchemy
pandas
newsapi-python
pytz
aiohttp
psutil==5.9.5